+++ shape/osci_CircleArc.h
@@ -6,16 +6,16 @@
 namespace osci {
 class CircleArc : public Shape {
 public:
-	CircleArc(double x, double y, double radiusX, double radiusY, double startAngle, double endAngle);
+	CircleArc(float x, float y, float radiusX, float radiusY, float startAngle, float endAngle);
 
-	Point nextVector(double drawingProgress) override;
-	void scale(double x, double y, double z) override;
-	void translate(double x, double y, double z) override;
-	double length() override;
+	Point nextVector(float drawingProgress) override;
+	void scale(float x, float y, float z) override;
+	void translate(float x, float y, float z) override;
+	float length() override;
 	std::unique_ptr<Shape> clone() override;
 	std::string type() override;
 
-	double x, y, radiusX, radiusY, startAngle, endAngle;
+	float x, y, radiusX, radiusY, startAngle, endAngle;
 	
 };
 } // namespace osci
