+++ shape/osci_Line.cpp
@@ -2,13 +2,13 @@
 
 namespace osci {
 
-Line::Line(double x1, double y1, double x2, double y2) : x1(x1), y1(y1), z1(0), x2(x2), y2(y2), z2(0) {}
+Line::Line(float x1, float y1, float x2, float y2) : x1(x1), y1(y1), z1(0), x2(x2), y2(y2), z2(0) {}
 
-Line::Line(double x1, double y1, double z1, double x2, double y2, double z2) : x1(x1), y1(y1), z1(z1), x2(x2), y2(y2), z2(z2) {}
+Line::Line(float x1, float y1, float z1, float x2, float y2, float z2) : x1(x1), y1(y1), z1(z1), x2(x2), y2(y2), z2(z2) {}
 
 Line::Line(Point p1, Point p2) : x1(p1.x), y1(p1.y), z1(p1.z), x2(p2.x), y2(p2.y), z2(p2.z) {}
 
-Point Line::nextVector(double drawingProgress) {
+Point Line::nextVector(float drawingProgress) {
 	return Point(
 		x1 + (x2 - x1) * drawingProgress,
 		y1 + (y2 - y1) * drawingProgress,
@@ -16,7 +16,7 @@ Point Line::nextVector(double drawingProgress) {
 	);
 }
 
-void Line::scale(double x, double y, double z) {
+void Line::scale(float x, float y, float z) {
 	x1 *= x;
 	y1 *= y;
 	z1 *= z;
@@ -25,7 +25,7 @@ void Line::scale(double x, double y, double z) {
 	z2 *= z;
 }
 
-void Line::translate(double x, double y, double z) {
+void Line::translate(float x, float y, float z) {
 	x1 += x;
 	y1 += y;
 	z1 += z;
@@ -34,11 +34,11 @@ void Line::translate(double x, double y, double z) {
 	z2 += z;
 }
 
-double Line::length(double x1, double y1, double z1, double x2, double y2, double z2) {
+float Line::length(float x1, float y1, float z1, float x2, float y2, float z2) {
 	return sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2) + pow(z2 - z1, 2));
 }
 
-double inline Line::length() {
+float inline Line::length() {
 	if (len < 0) {
 		len = length(x1, y1, z1, x2, y2, z2);
 	}
