+++ shape/osci_Point.h
@@ -7,22 +7,22 @@
 namespace osci {
 class Point : public Shape {
 public:
-    Point(double x, double y, double z);
-    Point(double x, double y);
-    Point(double val);
+    Point(float x, float y, float z);
+    Point(float x, float y);
+    Point(float val);
     Point();
 
-    Point nextVector(double drawingProgress) override;
-    void scale(double x, double y, double z) override;
-    void translate(double x, double y, double z) override;
-    double length() override;
-    double magnitude();
+    Point nextVector(float drawingProgress) override;
+    void scale(float x, float y, float z) override;
+    void translate(float x, float y, float z) override;
+    float length() override;
+    float magnitude();
     std::unique_ptr<Shape> clone() override;
     std::string type() override;
 
-    void rotate(double rotateX, double rotateY, double rotateZ);
+    void rotate(float rotateX, float rotateY, float rotateZ);
     void normalize();
-    double innerProduct(Point& other);
+    float innerProduct(Point& other);
 
     std::string toString();
 
@@ -30,29 +30,29 @@ public:
     bool operator==(const Point& other);
     bool operator!=(const Point& other);
     Point operator+(const Point& other);
-    Point operator+(double scalar);
-    friend Point operator+(double scalar, const Point& point);
+    Point operator+(float scalar);
+    friend Point operator+(float scalar, const Point& point);
     Point operator-(const Point& other);
     Point operator-();
-    Point operator-(double scalar);
-    friend Point operator-(double scalar, const Point& point);
+    Point operator-(float scalar);
+    friend Point operator-(float scalar, const Point& point);
     Point operator*(const Point& other);
-    Point operator*(double scalar);
-    friend Point operator*(double scalar, const Point& point);
-    Point operator/(double scalar);
-    friend Point operator/(double scalar, const Point& point);
+    Point operator*(float scalar);
+    friend Point operator*(float scalar, const Point& point);
+    Point operator/(float scalar);
+    friend Point operator/(float scalar, const Point& point);
 
     // Compound assignment operators
     Point& operator+=(const Point& other);
-    Point& operator+=(double scalar);
+    Point& operator+=(float scalar);
     Point& operator-=(const Point& other);
-    Point& operator-=(double scalar);
+    Point& operator-=(float scalar);
     Point& operator*=(const Point& other);
-    Point& operator*=(double scalar);
-    Point& operator/=(double scalar);
+    Point& operator*=(float scalar);
+    Point& operator/=(float scalar);
 
-    double x, y, z;
+    float x, y, z;
     
-    static double EPSILON;
+    static float EPSILON;
 };
 } // namespace osci
