+++ shape/osci_QuadraticBezierCurve.h
@@ -6,6 +6,6 @@
 namespace osci {
 class QuadraticBezierCurve : public CubicBezierCurve {
 public:
-	QuadraticBezierCurve(double x1, double y1, double x2, double y2, double x3, double y3);
+	QuadraticBezierCurve(float x1, float y1, float x2, float y2, float x3, float y3);
 };
 } // namespace osci
