+++ shape/osci_Shape.cpp
@@ -4,16 +4,16 @@
 
 namespace osci {
 
-double Shape::totalLength(std::vector<std::unique_ptr<Shape>>& shapes) {
-    double length = 0.0;
+float Shape::totalLength(std::vector<std::unique_ptr<Shape>>& shapes) {
+    float length = 0.0;
 	for (auto& shape : shapes) {
 		length += shape->length();
 	}
 	return length;
 }
 
-void Shape::normalize(std::vector<std::unique_ptr<Shape>>& shapes, double width, double height) {
-    double maxDim = std::max(width, height);
+void Shape::normalize(std::vector<std::unique_ptr<Shape>>& shapes, float width, float height) {
+    float maxDim = std::max(width, height);
     
 	for (auto& shape : shapes) {
         shape->scale(2.0 / maxDim, -2.0 / maxDim, 2.0 / maxDim);
@@ -24,25 +24,25 @@ void Shape::normalize(std::vector<std::unique_ptr<Shape>>& shapes, double width,
 }
 
 void Shape::normalize(std::vector<std::unique_ptr<Shape>>& shapes) {
-    double oldHeight = height(shapes);
-	double oldWidth = width(shapes);
-	double maxDim = std::max(oldHeight, oldWidth);
+    float oldHeight = height(shapes);
+	float oldWidth = width(shapes);
+	float maxDim = std::max(oldHeight, oldWidth);
 
 	for (auto& shape : shapes) {
 		shape->scale(2.0 / maxDim, -2.0 / maxDim, 2.0 / maxDim);
 	}
 
 	Point max = maxVector(shapes);
-	double newHeight = height(shapes);
+	float newHeight = height(shapes);
 
 	for (auto& shape : shapes) {
 		shape->translate(-1.0, -max.y + newHeight / 2.0, 0.0);
 	}
 }
 
-double Shape::height(std::vector<std::unique_ptr<Shape>>& shapes) {
-	double maxY = std::numeric_limits<double>::min();
-	double minY = std::numeric_limits<double>::max();
+float Shape::height(std::vector<std::unique_ptr<Shape>>& shapes) {
+	float maxY = std::numeric_limits<float>::min();
+	float minY = std::numeric_limits<float>::max();
 
     Point vectors[4];
 
@@ -60,9 +60,9 @@ double Shape::height(std::vector<std::unique_ptr<Shape>>& shapes) {
     return std::abs(maxY - minY);
 }
 
-double Shape::width(std::vector<std::unique_ptr<Shape>>& shapes) {
-    double maxX = std::numeric_limits<double>::min();
-    double minX = std::numeric_limits<double>::max();
+float Shape::width(std::vector<std::unique_ptr<Shape>>& shapes) {
+    float maxX = std::numeric_limits<float>::min();
+    float minX = std::numeric_limits<float>::max();
 
     Point vectors[4];
 
@@ -81,15 +81,15 @@ double Shape::width(std::vector<std::unique_ptr<Shape>>& shapes) {
 }
 
 Point Shape::maxVector(std::vector<std::unique_ptr<Shape>>& shapes) {
-    double maxX = std::numeric_limits<double>::min();
-    double maxY = std::numeric_limits<double>::min();
+    float maxX = std::numeric_limits<float>::min();
+    float maxY = std::numeric_limits<float>::min();
 
     for (auto& shape : shapes) {
         Point startVector = shape->nextVector(0);
         Point endVector = shape->nextVector(1);
 
-        double x = std::max(startVector.x, endVector.x);
-        double y = std::max(startVector.y, endVector.y);
+        float x = std::max(startVector.x, endVector.x);
+        float y = std::max(startVector.y, endVector.y);
 
         maxX = std::max(x, maxX);
         maxY = std::max(y, maxY);
@@ -109,8 +109,8 @@ void Shape::removeOutOfBounds(std::vector<std::unique_ptr<Shape>>& shapes) {
         if ((start.x < 1 && start.x > -1) || (start.y < 1 && start.y > -1)) {
             if ((end.x < 1 && end.x > -1) || (end.y < 1 && end.y > -1)) {
 				if (shapes[i]->type() == "Line") {
-                    Point newStart(std::min(std::max(start.x, -1.0), 1.0), std::min(std::max(start.y, -1.0), 1.0));
-                    Point newEnd(std::min(std::max(end.x, -1.0), 1.0), std::min(std::max(end.y, -1.0), 1.0));
+                    Point newStart(std::min(std::max(start.x, -1.0f), 1.0f), std::min(std::max(start.y, -1.0f), 1.0f));
+                    Point newEnd(std::min(std::max(end.x, -1.0f), 1.0f), std::min(std::max(end.y, -1.0f), 1.0f));
 					shapes[i] = std::make_unique<Line>(newStart.x, newStart.y, newEnd.x, newEnd.y);
                 }
                 keep = true;
