+++ shape/osci_Shape.h
@@ -10,23 +10,23 @@ namespace osci {
 class Point;
 class Shape {
 public:
-	virtual Point nextVector(double drawingProgress) = 0;
-	virtual void scale(double x, double y, double z) = 0;
-	virtual void translate(double x, double y, double z) = 0;
-	virtual double length() = 0;
+	virtual Point nextVector(float drawingProgress) = 0;
+	virtual void scale(float x, float y, float z) = 0;
+	virtual void translate(float x, float y, float z) = 0;
+	virtual float length() = 0;
 	virtual std::unique_ptr<Shape> clone() = 0;
 	virtual std::string type() = 0;
 
-	static double totalLength(std::vector<std::unique_ptr<Shape>>&);
-	static void normalize(std::vector<std::unique_ptr<Shape>>&, double, double);
+	static float totalLength(std::vector<std::unique_ptr<Shape>>&);
+	static void normalize(std::vector<std::unique_ptr<Shape>>&, float, float);
 	static void normalize(std::vector<std::unique_ptr<Shape>>&);
-	static double height(std::vector<std::unique_ptr<Shape>>&);
-	static double width(std::vector<std::unique_ptr<Shape>>&);
+	static float height(std::vector<std::unique_ptr<Shape>>&);
+	static float width(std::vector<std::unique_ptr<Shape>>&);
 	static Point maxVector(std::vector<std::unique_ptr<Shape>>&);
 	static void removeOutOfBounds(std::vector<std::unique_ptr<Shape>>&);
 
-	const double INVALID_LENGTH = -1.0;
+	const float INVALID_LENGTH = -1.0;
 
-	double len = INVALID_LENGTH;
+	float len = INVALID_LENGTH;
 };
 } // namespace osci
